+++ src/iblai_infra/__init__.py
@@ -1,3 +1,3 @@
 """ibl.ai Infrastructure Provisioning Tool."""
 
-__version__ = "1.14.0"
+__version__ = "1.15.0"